uncommited changes - untested!

+++ src/filters/merge.h
@@ -53,7 +53,7 @@ class MergeChroma : public GenericVideoFilter
  **/
 {
 public:
-  MergeChroma(PClip _child, PClip _clip, float _weight, int _test, IScriptEnvironment* env);  
+  MergeChroma(PClip _child, PClip _clip, double _weight, int _test, IScriptEnvironment* env);  
   PVideoFrame __stdcall GetFrame(int n, IScriptEnvironment* env);
 
   static AVSValue __cdecl Create(AVSValue args, void* user_data, IScriptEnvironment* env);
@@ -71,7 +71,7 @@ class MergeLuma : public GenericVideoFilter
  **/
 {
 public:
-  MergeLuma(PClip _child, PClip _clip, float _weight, int _test, IScriptEnvironment* env);  
+  MergeLuma(PClip _child, PClip _clip, double _weight, int _test, IScriptEnvironment* env);  
   PVideoFrame __stdcall GetFrame(int n, IScriptEnvironment* env);
 
   static AVSValue __cdecl Create(AVSValue args, void* user_data, IScriptEnvironment* env);
@@ -89,7 +89,7 @@ class MergeAll : public GenericVideoFilter
  **/
 {
 public:
-  MergeAll(PClip _child, PClip _clip, float _weight, int _test, IScriptEnvironment* env);  
+  MergeAll(PClip _child, PClip _clip, double _weight, int _test, IScriptEnvironment* env);  
   PVideoFrame __stdcall GetFrame(int n, IScriptEnvironment* env);
 
   static AVSValue __cdecl Create(AVSValue args, void* user_data, IScriptEnvironment* env);
